+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Linked track names with the tracks list. So whenever you change the content from owned or not, it reflects the changes in the other pages as well.
 
+### Changed
+
+- Moved "CONTENT" page to be the first one.
+
 ## [0.1.0 - alpha] - 2020-03-10 (2020 Season 2)
 
 ### Added